refactor(vite): group plugins into twig and assets subsystems
