+++ config/vite/plugins/assets/copy-src-assets.js
@@ -1,8 +1,8 @@
 /**
- * @file Twig template and component metadata copy plugin.
+ * @file Static source asset copy plugin.
  *
- * Copies canonical source Twig files and component metadata to the emitted dist
- * structure using the same routing rules as compiled JS and CSS entries.
+ * Copies non-code source assets beside the JS/CSS/Twig output that references
+ * them, preserving component and global routing semantics.
  */
 
 import { copyFileSync, mkdirSync } from 'fs';
@@ -11,23 +11,20 @@ import { dirname, join } from 'path';
 import {
   copiedComponentOutputPath,
   copiedGlobalOutputPath,
-} from '../project-structure.js';
+  findSourceRoot,
+} from '../../project-structure.js';
 import {
   createSourceFileIndex,
-  isComponentMetadataFile,
+  isStaticSourceAsset,
 } from './source-file-index.js';
 
-/** Determine whether a Twig file is a partial (filename starts with `_`). */
-const isPartial = (filePath) =>
-  (filePath.split('/')?.pop() || '').trim().startsWith('_');
-
 /**
- * Copy Twig templates and component metadata to `dist/`.
+ * Copy non-code assets from source roots to `dist/`.
  *
  * @param {{ structure: object, sourceFileIndex?: object }} opts - Plugin options.
  * @returns {import('vite').PluginOption} Copy plugin.
  */
-export function copyTwigFilesPlugin({
+export function copyAllSrcAssetsPlugin({
   structure,
   sourceFileIndex = createSourceFileIndex(structure),
 }) {
@@ -45,11 +42,11 @@ export function copyTwigFilesPlugin({
   };
 
   return {
-    name: 'emulsify-copy-twig-files',
+    name: 'emulsify-copy-all-src-assets',
     apply: 'build',
     enforce: 'post',
 
-    /** Capture the final outDir. */
+    /** Capture outDir. */
     configResolved(cfg) {
       outDir = cfg.build?.outDir || 'dist';
     },
@@ -57,23 +54,18 @@ export function copyTwigFilesPlugin({
     /** Copy before the mirror plugin moves dist/components to the project root. */
     writeBundle() {
       for (const file of sourceFileIndex.componentFiles()) {
-        if (file.absPath.endsWith('.twig')) {
-          if (isPartial(file.relPath)) continue;
-          copyToOutDir(
-            file.absPath,
-            copiedComponentOutputPath(file.absPath, structure),
-          );
-        } else if (isComponentMetadataFile(file.absPath)) {
-          copyToOutDir(
-            file.absPath,
-            copiedComponentOutputPath(file.absPath, structure),
-          );
-        }
+        if (!isStaticSourceAsset(file.absPath)) continue;
+        copyToOutDir(
+          file.absPath,
+          copiedComponentOutputPath(file.absPath, structure),
+        );
       }
 
       for (const file of sourceFileIndex.globalFiles()) {
-        if (!file.absPath.endsWith('.twig')) continue;
-        if (isPartial(file.relPath)) continue;
+        if (!isStaticSourceAsset(file.absPath)) continue;
+        if (findSourceRoot(file.absPath, structure.componentRootRecords)) {
+          continue;
+        }
         copyToOutDir(
           file.absPath,
           copiedGlobalOutputPath(file.absPath, structure),
