+++ config/vite/plugins/twig/extension-installers.test.js
@@ -5,7 +5,7 @@
 import {
   VIRTUAL_TWIG_EXTENSION_INSTALLERS_ID,
   virtualTwigExtensionInstallersPlugin,
-} from './twig-extension-installers.js';
+} from './extension-installers.js';
 
 describe('virtual Twig extension installer module plugin', () => {
   it('resolves and loads the virtual module', () => {
