+++ config/vite/plugins/twig/virtual-twig-asset-sources.js
@@ -7,11 +7,11 @@
 
 import { readdirSync } from 'fs';
 import { relative, resolve } from 'path';
-import { safeExists } from '../utils/fs-safe.js';
-import { toPosixPath } from '../utils/paths.js';
-import { unique } from '../../../src/extensions/shared/lists.js';
-import { toRootRelativePath } from '../../../src/extensions/shared/root-relative.js';
-import { INLINE_ASSET_EXTS } from '../../../src/storybook/twig/source-extensions.js';
+import { safeExists } from '../../utils/fs-safe.js';
+import { toPosixPath } from '../../utils/paths.js';
+import { unique } from '../../../../src/extensions/shared/lists.js';
+import { toRootRelativePath } from '../../../../src/extensions/shared/root-relative.js';
+import { INLINE_ASSET_EXTS } from '../../../../src/storybook/twig/source-extensions.js';
 
 export const VIRTUAL_TWIG_ASSET_SOURCES_ID =
   'virtual:emulsify-twig-asset-sources';
@@ -247,7 +247,7 @@ export function generateVirtualTwigAssetSourcesModule(env) {
     .join(',\n');
 
   return `/**
- * Virtual module generated by config/vite/plugins/virtual-twig-asset-sources.js.
+ * Virtual module generated by config/vite/plugins/twig/virtual-twig-asset-sources.js.
  *
  * Raw text assets stay lazy and load only when Twig source() requests them.
  */
