+++ config/vite/plugins/twig/virtual-twig-globs.js
@@ -5,8 +5,8 @@
  * virtual module instead of replacing placeholder strings in resolver source.
  */
 
-import { unique } from '../../../src/extensions/shared/lists.js';
-import { toRootRelativePath } from '../../../src/extensions/shared/root-relative.js';
+import { unique } from '../../../../src/extensions/shared/lists.js';
+import { toRootRelativePath } from '../../../../src/extensions/shared/root-relative.js';
 
 export const VIRTUAL_TWIG_GLOBS_ID = 'virtual:emulsify-twig-globs';
 const RESOLVED_VIRTUAL_TWIG_GLOBS_ID = `\0${VIRTUAL_TWIG_GLOBS_ID}`;
@@ -56,7 +56,7 @@ export function generateVirtualTwigGlobsModule(env) {
     : '';
 
   return `/**
- * Virtual module generated by config/vite/plugins/virtual-twig-globs.js.
+ * Virtual module generated by config/vite/plugins/twig/virtual-twig-globs.js.
  *
  * Compiled Twig modules stay eager for synchronous story rendering. Raw source
  * entries stay lazy and load only when Twig source() requests a template.
