+++ config/vite/plugins/twig/vituum-patch.js
@@ -8,9 +8,9 @@
 import twig from '@vituum/vite-plugin-twig';
 import Twig from 'twig';
 
-import { registerTwigExtensions } from '../../../src/extensions/twig/index.js';
+import { registerTwigExtensions } from '../../../../src/extensions/twig/index.js';
 import { makeTwigPluginOptions } from './twig-module.js';
-import { registerConfiguredTwigExtensions } from '../twig-extensions.js';
+import { installProjectTwigExtensions } from './extensions.js';
 
 const EXPECTED_PLUGIN_NAMES = [
   '@vituum/vite-plugin-core:bundle',
@@ -140,7 +140,7 @@ function stripExpectedHooks(pluginOption) {
  */
 export function makeTwigPlugins(env, options = makeTwigPluginOptions(env)) {
   registerTwigExtensions(Twig);
-  registerConfiguredTwigExtensions(Twig, options);
+  installProjectTwigExtensions(Twig, options);
 
   const twigPlugins = twig(options);
   const normalizedPlugins = Array.isArray(twigPlugins)
