+++ package.json
@@ -55,25 +55,25 @@
     "config/vite/entries.js",
     "config/vite/environment.js",
     "config/vite/platforms.js",
-    "config/vite/plugins/copy-src-assets.js",
-    "config/vite/plugins/copy-twig-files.js",
-    "config/vite/plugins/css-asset-relativizer.js",
+    "config/vite/plugins/assets/copy-src-assets.js",
+    "config/vite/plugins/assets/copy-twig-files.js",
+    "config/vite/plugins/assets/css-asset-relativizer.js",
+    "config/vite/plugins/assets/mirror-components.js",
+    "config/vite/plugins/assets/source-file-index.js",
+    "config/vite/plugins/assets/svg-sprite.js",
     "config/vite/plugins/index.js",
-    "config/vite/plugins/mirror-components.js",
     "config/vite/plugins/require-context.js",
-    "config/vite/plugins/source-file-index.js",
-    "config/vite/plugins/svg-sprite.js",
-    "config/vite/plugins/twig-extension-installers.js",
-    "config/vite/plugins/twig-module.js",
-    "config/vite/plugins/virtual-twig-asset-sources.js",
-    "config/vite/plugins/virtual-twig-globs.js",
-    "config/vite/plugins/vituum-patch.js",
+    "config/vite/plugins/twig/extension-installers.js",
+    "config/vite/plugins/twig/extensions.js",
+    "config/vite/plugins/twig/twig-module.js",
+    "config/vite/plugins/twig/virtual-twig-asset-sources.js",
+    "config/vite/plugins/twig/virtual-twig-globs.js",
+    "config/vite/plugins/twig/vituum-patch.js",
     "config/vite/plugins/yaml-module.js",
     "config/vite/plugins.js",
     "config/vite/project-config.js",
     "config/vite/project-extensions.js",
     "config/vite/project-structure.js",
-    "config/vite/twig-extensions.js",
     "config/vite/utils/fs-safe.js",
     "config/vite/utils/paths.js",
     "config/vite/utils/react-singleton.js",
